+++ pkg/ffuf/config.go
@@ -52,6 +52,7 @@ type Config struct {
 	StopOn403               bool                  `json:"stop_403"`
 	StopOnAll               bool                  `json:"stop_all"`
 	StopOnErrors            bool                  `json:"stop_errors"`
+	StopAfterN		int		      `json:"stop_after_n"`
 	Threads                 int                   `json:"threads"`
 	Timeout                 int                   `json:"timeout"`
 	Url                     string                `json:"url"`
@@ -103,6 +104,7 @@ func NewConfig(ctx context.Context, cancel context.CancelFunc) Config {
 	conf.StopOn403 = false
 	conf.StopOnAll = false
 	conf.StopOnErrors = false
+	conf.StopAfterN = 0 // 0 implies we do not stop
 	conf.Timeout = 10
 	conf.Url = ""
 	conf.Verbose = false